function/requirements.txt
azure-functions
onnxruntime
Pillow
